Metadata cleaners added
I had the marvelous opportunity to be asked (yet again) to actually do something at my job. All the files in the system had the metadata untouched but they had to be sent to a client, instead of removing one by one the metadata of each file, I did a metadata cleaner and reader to check if there was any left.
It may or may not remove esssential information of your files, i.e. removing titles, so use it accordingly.